Make PWM input to 16bits as sometimes the input larger than 8bits. This avoid some weird behavior somtimes

+++ src/board/motor/motor.h
@@ -7,7 +7,7 @@ extern "C" {
 #endif
 
 void motor_init(void);
-void motor_start(int8_t left_speed, int8_t right_speed);
+void motor_start(int16_t left_speed, int16_t right_speed);
 void motor_stop(void);
 
 #ifdef  __cplusplus
